Reference

This is the reference documentation for the ParticleHolography package.

Index

ParticleHolography.cu_get_reconst_vol

ParticleHolography.cu_get_reconst_vol_and_xyprojection

ParticleHolography.cu_get_reconst_xyprojection

ParticleHolography.cu_phase_retrieval_holo

ParticleHolography.cu_transfer

ParticleHolography.cu_transfer_sqrt_arr

ParticleHolography.cvgray2floatimg

ParticleHolography.find_external_contours

ParticleHolography.floatimg2cvgray

ParticleHolography.getErrorVec

ParticleHolography.getYacobian

ParticleHolography.get_distortion_coefficients

ParticleHolography.load_gray2float

ParticleHolography.modified_Cholesky_decomposition

ParticleHolography.quadratic_distortion_correction

ParticleHolography.simultanious_equation_solver

Functions

ParticleHolography.cu_get_reconst_vol — Method

cu_get_reconst_vol(holo, transfer_front, transfer_dz, slices)

Reconstruct the observation volume from the light field light_field using the transfer functions transfer_front and transfer_dz. transfer_front propagates the light field to the front of the volume, and transfer_dz propagates the light field between the slices. slices is the number of slices in the volume.

Arguments

light_field::CuArray{ComplexF32,2}: The light_field to reconstruct. In Gabor's holography, this is the square root of the hologram.

transfer_front::CuArray{ComplexF32,2}: The transfer function to propagate the light field to the front of the volume.

transfer_dz::CuArray{ComplexF32,2}: The transfer function to propagate the light field between the slices.

slices::Int: The number of slices in the volume.

Returns

CuArray{Float32,3}: The reconstructed volume.

ParticleHolography.cu_get_reconst_vol_and_xyprojection — Method

cu_get_reconst_vol_and_xyprojection(light_field, transfer_front, transfer_dz, slices)

Reconstruct the observation volume from the light field light_field and get the XY projection of the volume using the transfer functions transfer_front and transfer_dz. transfer_front propagates the light field to the front of the volume, and transfer_dz propagates the light field between the slices. slices is the number of slices in the volume.

Arguments

light_field::CuArray{ComplexF32,2}: The light_field to reconstruct. In Gabor's holography, this is the square root of the hologram.

transfer_front::CuArray{ComplexF32,2}: The transfer function to propagate the light field to the front of the volume.

transfer_dz::CuArray{ComplexF32,2}: The transfer function to propagate the light field between the slices.

slices::Int: The number of slices in the volume.

Returns

CuArray{Float32,3}: The reconstructed volume.

CuArray{Float32,2}: The XY projection of the reconstructed volume.

ParticleHolography.cu_get_reconst_xyprojection — Method

cu_get_reconst_xyprojectin(light_field, transfer_front, transfer_dz, slices)

Get the XY projection of the reconstructed volume from the light field light_field using the transfer functions transfer_front and transfer_dz. transfer_front propagates the light field to the front of the volume, and transfer_dz propagates the light field between the slices. slices is the number of slices in the volume.

Arguments

light_field::CuArray{ComplexF32,2}: The light_field to reconstruct. In Gabor's holography, this is the square root of the hologram.

transfer_front::CuArray{ComplexF32,2}: The transfer function to propagate the light field to the front of the volume.

transfer_dz::CuArray{ComplexF32,2}: The transfer function to propagate the light field between the slices.

slices::Int: The number of slices in the volume.

Returns

CuArray{Float32,2}: The XY projection of the reconstructed volume.

ParticleHolography.cu_phase_retrieval_holo — Method

cu_phase_retrieval_holo(holo1, holo2, transfer, invtransfer, priter, datlen)

Perform the Gerchberg-Saxton algorithm-based phase retrieving on two holograms and return the retrieved light field at the z-coordinate point of holo1. The algorithm is repeated priter times. holo1 and holo2 are the holograms (I = |phi|^2) of the object at two different z-coordinates. transfer and invtransfer are the transfer functions for the propagation from holo1 to holo2 and vice versa. datlen is the size of the holograms.

Arguments

holo1::CuArray{Float32,2}: The hologram at the z-cordinate of closer to the object.

holo2::CuArray{Float32,2}: The hologram at the z-coordinate of further from the object.

transfer::CuArray{ComplexF32,2}: The transfer function from holo1 to holo2.

invtransfer::CuArray{ComplexF32,2}: The transfer function from holo2 to holo1.

priter::Int: The number of iterations to perform the algorithm.

datlen::Int: The size of the holograms.

Returns

CuArray{ComplexF32,2}: The retrieved light field at the z-coordinate of holo1.

ParticleHolography.cu_transfer — Method

cu_transfer(z0, datLen, wavLen, d_sqr)

Create a CuArray of size datLen x datLen with the values of the transfer function for a given propagated distance z0. d_sqr can be obtained with cutransfersqrtarr(datlen, wavlen, dx).

Arguments

z0::AbstractFloat: The distance to propagate the wave.

datLen::Int: The size of the CuArray.

wavLen::AbstractFloat: The wavelength of the light.

d_sqr::CuArray{Float32,2}: The square of the distance from the center of the hologram, obtained with cutransfersqrtarr(datlen, wavlen, dx).

Returns

CuArray{ComplexF32,2}: The transfer function for the propagation.

ParticleHolography.cu_transfer_sqrt_arr — Method

cu_transfer_sqrt_arr(datlen, wavlen, dx)

Create a CuArray of size datlen x datlen with the values of the square-root part of the transfer function.

Arguments

datlen::Int: The size of the CuArray.

wavlen::AbstractFloat: The wavelength of the light.

dx::AbstractFloat: The pixel size of the hologram.

Returns

CuArray{Float32,2}: The square-root part of the transfer function.

ParticleHolography.cvgray2floatimg — Method

cvgray2floatimg(img)

Convert a OpenCV capable image to a Array{Float32, 2} image.

Arguments

img::Array{N0f8, 3}: The image to convert.

Returns

Array{Float32, 2}: The image as a Float32 array.

ParticleHolography.find_external_contours — Method

find_external_contours(image)

Finds non-hole contours in binary images. Equivalent to CVRETREXTERNAL and CVCHAINAPPROX_NONE modes of the findContours() function provided in OpenCV.

Arguments

image: The binary image.

Returns

Vector{Vector{CartesianIndex}}: A vector of contours. Each contour is a vector of CartesianIndex.

ParticleHolography.floatimg2cvgray — Method

floatimg2cvgray(img)

Convert a Array{Float32, 2} image to a OpenCV capable image.

Arguments

img::Array{Float32, 2}: The image to convert.

Returns

Array{N0f8, 3}: The image as a OpenCV capable image.

ParticleHolography.getErrorVec — Function

getErrorVec(vecMap, coefa, gridSize = 128, imgSize = 1024)

Return the vector `errorVec` obtained by calculating the difference between `targetX` `targetY` and `procX` `procY` obtained by converting `targetX` `targetY` from the `gridx` `gridy` set in the first image to the second image. Also returns `gridx` `gridy` for Jacobian acquisition.

Arguments

vecMap: The PIV map.

coefa: The coefficients for the quadratic distortion correction.

gridSize: The size of the grid.

imgSize: The size of the image.

ParticleHolography.getYacobian — Function

getYacobian(imgSize = 1024, gridSize = 128)

Get the Jacobian matrix for the quadratic distortion correction. The default image size is 1024, and the grid size is 128.

Arguments

imgSize::Int: The size of the image.

gridSize::Int: The size of the grid.

Returns

Array{Float64,2}: The Jacobian matrix.

ParticleHolography.get_distortion_coefficients — Method

get_distortion_coeficients(img1, img2, gridSize = 128, intrSize = 128, srchSize = 256)

Get the distortion coefficients from the two images img1 and img2. The default grid size is 128, the default search size is 256, and the default image size is 128.

Arguments

img1::Array{<:AbstractFloat,2}: The first image.

img2::Array{<:AbstractFloat,2}: The second image.

gridSize::Int: The size of the grid.

intrSize::Int: The size of the search.

srchSize::Int: The size of the search.

Returns

Vector{<:AbstractFloat}: The distortion coefficients.

ParticleHolography.load_gray2float — Method

load_gray2float(path)

Load a grayscale image from a file and return it as a Array{Float32, 2} array.

Arguments

path::String: The path to the image file.

Returns

Array{Float32, 2}: The image as a Float32 array.

ParticleHolography.modified_Cholesky_decomposition — Method

modified_Cholesky_decomposition(A)

This function decomposes the target matrix A to A = LDL^T. The diagonal components of the return matrix are the inverse of D, and the lower left components correspond to L. 対象行列Aを A = LDL^T に変換します。戻り値行列の対角成分は D の逆数、左下成分は L に一致します。

Arguments

A::Array{<:AbstractFloat,2}: The target matrix.

Returns

Array{AbstractFloat,2}: The decomposed matrix.

ParticleHolography.quadratic_distortion_correction — Method

quadratic_distortion_correction(img, coefa)

Correct the quadratic distortion in the grayscale image img using the coefficients coefa. img have to be square, and coefa have to be a vector of 12 coefficients.

Arguments

img::Array{<:AbstractFloat,2}: The image to correct.

coefa::Vector{<:AbstractFloat}: The coefficients to correct the distortion.

Returns

Array{AbstractFloat,2}: The corrected image.

ParticleHolography.simultanious_equation_solver — Method

simultanious_equation_solver(chlskyMat,yacob,errorArray)

Solve the equation chlskyMat x = - yacob errorArray using the Cholesky decomposition matrix chlskyMat, Jacobian yacob, and error vector errorArray.

Arguments

chlskyMat::Array{<:AbstractFloat,2}: The Cholesky decomposition matrix.

yacob::Array{<:AbstractFloat,2}: The Jacobian matrix.

errorArray::Array{<:AbstractFloat,1}: The error vector.

Returns

Array{AbstractFloat,1}: The solution vector.